refactor: Rename the async module to aio
`async` is a reserved keyword in rust 2018 so this avoids the need to
write `r#async` in it.

Suggested by @jwilm in https://github.com/mitsuhiko/redis-rs/issues/180

Closes https://github.com/mitsuhiko/redis-rs/issues/180

BREAKING CHANGE

Rename uses of the `async` module to `aio`
